Added logic to skip migrations for sqlite databases, as sqlite does not support altering schema data.
Added non-py Alembic files to packaging process, as they must be present for
migration sprocessing to success in deployed environments.

Modified the version stamping script to stamp versions without need for templated fields to
be preserved, allowing for repeated calls to update scripts automatically. This is
especially useful for development purposes.

Change-Id: I98af79fee8fec3ac131e54035d00ce0a8f855a63