Restrict use of 'plain_text' to 'text/plain' MIME
When creating a secret with a non 'text/plain' MIME type with
supplied 'plain_text', a ValueError would be thrown during the
encryption step. This resulted in a 500 HTTP status. Now when
a create secret request is sent with a 'plain_text' field
provided and a non 'text/plain' MIME type, a 400 HTTP status
is returned.

Fixes: bug #1200659
Change-Id: I9df8ec52c96c6cbe63c1f1af85e9a60ef507ff69