Fix order creation bug involving content types
Order creation was trying to look up mime_type, which is no longer
a supported field. MIME type was replaced with payload_content_type
and a secret_payload_content_type column was created. The database
must be flushed or the addition must be migrated for this to work
properly. Also require payload_content_type when creating orders.

Fixes: bug #1208528
Change-Id: I51e93ec63ce2786ddb7a47d528cc70507a7d88dd