Refine the Secret metadata response by *not* returning the 'content-encodings'.
The 'content-encodings' for Secrets define the types of transfer protocols available
to retrieve stored Secret data with. This really should be negotiated between the
client and server at the HTTP level, so there is no need for Barbican to provide
a list of such 'content-encodings' for clients to use. Thus clients will receive a
406 response if they specify an Accept-Encoding not supported by Barbican. The API
documentation will be updated to reflect this point.

Implements: blueprint api-secret-get-encodings-refine

Change-Id: Iaaa7c13e0a2103d39f5d8425ca08b9aee849417c