Remove base64 encoding feature from decrypted secrets GET API.
We are having issues supporting the base64 encoding feature when getting decrypted
secrets. In particular the Accept-Encoding http header appears to be overridden by
browsers such as Chrome and the 'requests' library used by Cloud Cafe, even though
'base64' is set for header. Rather than contend with this ill-supported header in the
long term, it is better to just remove this feature for the time being.

Change-Id: Iafafe364d273cd0ed6d5615729bdedce3c1335f5
Implements: blueprint api-secrets-remove-base64-support