Modify version template for the final OpenStack Havana release
The version template defines the prefix for each build version that does not
change automatically, and a final suffix that changes on each interim release
candidate built. This commit set that prefix to the Havana release prefix,
which required being able to accommodate versions such as this aa.bb.cc.dd
instead of just aa.bb.cc as before.

Change-Id: I983efb7f0c2407351d276c6ed80bfa0e98587828