Remove unneeded Barbican worker script from generated RPM.
An unneeded script (bin/barbican-worker) was being specified in the setup.py
file, but not specified in the RPM spec file as it is not needed for deployment.
This was leading to an 'unneeded file' error in the RPM build process. This
change removes that unneeded file from setup.py.

Change-Id: Ib9913ca152fc333f9e7ef9f1f749b2907970bd35