host_href should include scheme
Barbican should not presuppose that host_href will be an HTTP endpoint.
This does not work when SSL is being used (ie. in the case there
haproxy is performing SSL termination for Barbican)

This change simply requires that host_href include the scheme to be
used when providing secret and/or order refs.

Change-Id: I22f6b13b693801322716d4c43ce8d508eca114dd
Closes-Bug: #1258598