Modify python-dateutil RPM build process
The current build process for the python-dateutil RPM generates a
package that when installed lays a number of files on a filesystem
that only root can read.  This is problematic for non-root owned
processes that are trying to read these files.  Adding this patch
opens up those files to reading by process that aren't running as
root.  For example, this would allow for the running of the
reference uWSGI based implementation as a non-root user.

Fixes: bug #1288033
Change-Id: I3bff45bab272de888992cb42a8b8709a3fcd8ec6