Remove v1 from the falcon app route
This bug was introduced by a recent change
"Barbican should not do rbac on version api call"
Where the main barbican app was split up into two falcon apps,
one for the version resource, and another for the rest of the
calls.

In the paste file, the latter app was mounted on 'v1' but
because the routes in the falcon app also map to 'v1' this resulted
in all calls being routed to 'v1/v1/{tenant}/...'

This change removes the routes from the app, so that only the
'v1' in the paste configuration is mapped, which fixes the
double v1 route.

Change-Id: I05746f11d262ec7aa4f60aeb44e41217e8bc8a44
Closes-Bug: #1296969