Adds SecretMetadatum table
This patch is the first planned in a series to add KMIP support to Barbican.
Barbican currently returns metadata about secrets to the user, but adding
secret_metadata table will allow users to set additional metadata options
that are not supported by Barbican.

After this submission is accepted, another patch will be created to move the
current Secret attribute metadata to the new metadata table.

Change-Id: Ibc990f798b118eeda705360e40ecf439d5a1b0cd
Implements: blueprint add-metadata-to-secrets