Remove redundant code on offset, limit parameters.
The API resource modules for secrets, orders and containers had logic to clean up
the provided offset and limit parameters for paged GET requests. However, this logic
is already provided in the model repositories that the API modules call to make the
list query, hence this code can be removed from the API modules.

Change-Id: Icd6c29156f3bf80582984dbeb1d35f14fd13272f