Disable empty or null names for crypto plugins
Plugins used for crypto processing are loaded using stevedore, which in
turn uses the plugin_name attribute to load it. In this case, it makes
no sense to have names that could be None or invalid, thus I modified
the model to have this as an attribute. And, now the plugin repo class
doesn't allow the saving of plugins with an invalid plugin name.

On the other hand, it used to be the case that the generation of the
plugin names would return None as a string, instead of it's actual
value, which could potentially lead to adding erroneous entries in the
plugin store. This was fixed also.

Change-Id: Ic50f2df5fb0778644bbd6aea626fa165f2f972c0
Closes-Bug: #1325972