Fix data migration script error
There are two types of data migration:
1. Automatic - happens during barbican api startup
2. Manual - run on demand when migration is required

The Manual step is done by running the bin/barbican-db-manage.py
script. This script fails with an "UnboundExecutionError" due
to a conflict between SQLAlchemy synching vs Alembic version
processing in checking for the existence of the table. This commit
fixes the issue by replacing the SQLAlchemy metadata calls with
alembic calls by reusing the context/connection from the existing
"op".

Change-Id: I9bf65594a9e76b3f98d67bbd47a9cc7b97298de0
Closes-Bug: #1326654