Add dogtag plugin to new plugin structure
This patch adds the dogtag plugin (and tests) to the new plugin
structure under plugins/ and converts the plugin to an implementation
of the SecretStore interface.

The old dogtag_crypto_plugin will be retained until the top level
plugin manager and resources have been switched over, allowing us to
fully test the new dogtag plugin.

Change-Id: I20e2daca6dad2339d228e850a03868ac587089c0