Add code to retrieve secrets metadata and data with transport key
On GET request for secret metadata, if the plugin supports transport
key wrapping, then return the transport_key_ref with the metadata.
On a GET requests for the secret data, if a trans_wrapped_session_key
is provided, pass it to the plugin to use when encrypting the returned
secret.

Implements: blueprint add-wrapping-key-to-barbican-server
Change-Id: Ifdf2fce4c72419fd50cd490ae8515da2203b2c84