Eager load KEKDatum record when EncryptedDatum is retrieved
When trying to GET a decrypted secret from Barbican, an exception is
raised. This is due to the KEKDatum record associated with an
EncryptedDatum record attempting to be lazy-loaded after the SQLAlchemy
session is closed. SQLite does not raise an exception, but Postgres does.
Also, the lazy='joined' syntax did not make it obvious that an eager
load was called for, hence this syntax was changed to lazy=False with
a comment added.

Change-Id: I728e2052d8a916da5c1023daf8137774ac04f8f5
Closes-Bug: #1352020