Force uWSGI to set "Connection: close" header
uWSGI is currently not compliant with RFC 2616 section 14.10.
This will work around the problem by blindly applying the
correct header to responses, when using the default uWSGI
vassal configuration files. This does break persistent connections
in that case, so these configs should probably be changed for
production use (or uWSGI should not be used).

Change-Id: I56036768c46f1f727b10c7d231c3470dfa8213ea