Allow devstack to do git clone of barbican
Currently on F20 env when devstack tries to install barbican and
barbicanclient, it's failing as corresponding dir doesn't exist and
ERROR_ON_CLONE=True. As barbican is not part of devstack yet,
we don't expect this dir to be present in this case, we should
set ERROR_ON_CLONE=False to allow devstack to do git clone.

Related-Bug: #1353696
Change-Id: I3c362f1b2e6c9abf940ea68a12085dd747ede37c