Update versionbuild script to handle setup.cfg version
The versionbump script used to supply a monotonically increasing version for
rpm packages was modified to supply a 'version=' attribute to setup.cfg, to
override the pbr-generated version. Unfortunately, this change breaks devstack
deployments. So the version attribute was added back to setup.cfg. Now
versionbump functionality needs to be reverted back to adding the timestamp to
the hard-coded version stamp, but with the more Pythonic mods suggested in the
previous CR.

Closes-Bug: #1358455
Change-Id: Icc5a6a4319ea4517f2812168f25840253eb50e28