Initial connect orders resource to certificate processing
Connects the new orders type/meta API to the issue_certificate_request()
method of certificate worker tasking logic. The order plugin metadata is not
provided with this CR. Saving a certificate into a Container is also not
provided with this CR. This CR does address issues with the orders resource
processing.

Change-Id: Id5c6c06ecccd3d4d7eb27d974156dba06bd8fe6f
Implements: blueprint add-ssl-ca-support