Replace explicit assertion for function
In order to reduce the verbosity and repetition when checking the json
schema, there was a function written that does this: It asserts that the
given json input is compliant witht he schema, and throws a proper
exception if possible. Most of the instances of this pattern were
already replaced in the code, but it seems there was one missing (or was
added later). So it was rightly replaced.

Change-Id: Ic9af4ac71c9d91a2083fccd67ca3901859f427e5