Switch to running tests in parallel with testr
This commits switches the test runner in the functional tests to use
testr in parallel instead of nosetests. There are a few reasons for
doing this, firstly parallel testing ensures performs the functional
testing in a more realistic environment. Secondly, tempest which these
tests are based on doesn't support using nose, so to prevent any
potential issues we should use the same test runner here. Lastly, in
the future subunit output will be used to perform additionally
analysis on test runs. So if the same methods were to be used on the
functional tests they need to produce subunit streams from the test
runs.

Change-Id: I7daf2be58c44f2555861e47d3444523d75cc4e28