Remove restrictive hard-coded orders validation
The orders validation process for generating new secrets was enforcing
that parameters such as 'algorithm' were set to specific values, such
as 'aes'. This logic was put in place when only AES-generation was
provided by the default HSM plugin, as it is better to return a
synchronous error to clients than to wait until asynchronous
processing occurs. Now that 3rd party plugins are available that can
provide other generation types, this restriction is no longer valid.
So for Juno, the decision was to relax this validation logic, and to
let asynchronous processing by plugins flag any errors. In Kilo
however, the plan is to let the plugins perform this validation
up-front. TODO comments have been added to this effect.

Change-Id: I4e57c365ead698bf5435f0a5ca7be8076dfddb6a