Changes to get certificate issuance flow working
Multiple changes are needed to resolve database session
issues when accessing and updating order metadata, and
storing the resulting certificates and intermediates.

With this patch, we are able to place an order for a
certificate, get the certificate immediately issued from
a dogtag CA, and store the cert and intermediates
correctly.

Also added missing stanza in barbican-api.conf for
SecretStore

Closes-Bug: 1173533
Change-Id: I10785e1b6a3a73d4a9b1b2fcb1e6262e77a47b6a