Delete usage-indications from the model docstrings
I removed some lines from the docstrings in the models due to them
indicating the usage of the function in other parts of the source. This
is not useful as the code could be refactored later, or other usage
could be given to these functions, and thus the comments would become
easily obsolete. Thus, this commit removes them.

Change-Id: I467c8c0a2d7d149e18d2aff87cafb23896bbbca7