Fix 500 error when PUTing an order
Since the load_body function was not being used to parse the request
initially for orders, it was resulting in the exception that is seen
in the bug description. So, load_body was introduced. But, since
load_body uses a "read" for the ByteIO stream that contains the
request's payload, it could only be used once. So a "seek(0)" was
introduced to address this.

Change-Id: I66e9b2b6dbb7196cf5a639db52f9f25e82da1765
Closes-Bug: #1376438