Add plugin sub-folder and augment crypto plugin doc
This CR creates a folder for plugins, to make way for documents pages
for each of the plugin types used in Barbican. The existing crypto
plugin page is moved to this new structure and revamped to add more
information on how Barbican core works with the plugin. The new plugin
section's index page has generic information on plugins (extracted
from the original plugin.rst) as well as defining what Barbican 'core'
is and how it interacts with plugins. Dependent CRs to add content
for the other plugin types will be coming soon.

Change-Id: Ia63c4e46cc7f01c334b41ec22ce8e9373b1a79da