Add validation for certificate-order-api
This commit adds the basic framework for validation code needed
for the common cert API.  Note that there is a slight change to
the blueprint in that for the stored-key type, we expect to see
a container_ref rather than a public_key_ref, so that certain
validations (like access to the private key) can be done.

We expect subsequent commits to fill in the missing logic in the
functions and enable the relevant tests.  Also, once this process
is complete, it makes sense to refactor the Validator code along
the lines of bp/split-validator-module.

Implements: blueprint certificate-order-api
Change-Id: Ieffd0c6f8b97b9bb5b1d0d4cf9e8eea0a6277ca0