Actually run type order creation test
The running of the type order creation test used to be skipped due to
a pending CR. Now the test is ran, but there seemed to be a problem
where the proper project ID was not added to the calls. This was due
to the test not being properly set. Now this was fixed.

Change-Id: I64d00db8bc6f39d59a0fa42ad9af52bd1014a1f7