Plugin contract changes for the certificate-order-api
This patch adds the plugin contract changes and code needed
to pass the new cert request types to the plugins.

Note that this is different from the blueprint - in that we
now introduce a supported_request_types() method, and we no longer use separate
API calls to distinguish between types.

Missing logic in some functions is expected to be filled in
in subsequent patches.

Change-Id: If4810b5e96ade5255d7886eb9480157329380b97
Implements: blueprint certificate-order-api