Change keystone_id for external_id in model
The Project model contains a reference to its ID as assigned by an
external entity (usually keystone). Initially the name keystone_id was
given, but it was agreed that in the effort of standardizing
everything to use the v3 nomenclature (projects instead of tenants)
the wording 'keystone_id' would be replaced, as it was initially used
because of the aforementioned nomenclature confusion. Thus, to remove
redundancy and express what that ID actually is, the name is changed
in the Project model.

Partially implements: blueprint replace-concept-of-tenants-for-projects

Change-Id: I521f88af2c724af2a4e374169d76140829add54b