Replace instances of keystone_id from the code
Instead of using the word "keystone_id" to refer to the ID assigned to
a certain project from the outside, a more explicit keyword
"external_project id" is used; This was done so in order to reflect
that this is not necessarily the id given by Barbican to the Project
entity, but the one that we get from an external entity (usually
keystone). On the other hand, the simple "project_id" keyword was used
where the distinction is not so important to understand.

Change-Id: Ica61923af11d8e02b38737effd71ba5a7c3396a4
Implements: blueprint replace-concept-of-tenants-for-projects