Return the actual name value for entities
It used to be the case that for secrets and containers, if the name
was initially set to null, the GETing of the secret would actually
return the UUID, which is not what's reflected in the database. This
is now fixed. Issuing a GET for those entities will return what's
actually in the database.

On the other hand, there was a bug where the validator for secrets was
actually turning empty strings into None, and this would be saved in
the database. This was now fixed.

Change-Id: I8ebcf752e47be97ecdd4cb3c5ab152b6717b5341