Only de-serialize objects when possible
If the response obtained by the client in the functional tests was not
successful, the de-serialization into a model object will throw an
exception, since the response doesn't contain a JSON string that could
be used for creating such an object; It will contain the error code
and a relevant error message. This fixes that and also adds a little
bit more logging when de-serializing objects.

The reason for adding this is that, even though the exceptions that
this threw were caught at some point, it only generated unnecessary
noise in the logs.

Change-Id: I2fbf9ddbe21aaea3dcf112f4bff39942625599ff