Resolve intermittent HTTP 404 in devstack gate
An intermittent 404 would occur because of the way that
Tempest keystone v3 authprovider was resolving base URLs.
Barbican wasn't passing in a region code, and the tempest
v3 auth provider was picking the "first" entry in the
entrypoint list, which was often the Barbican admin URL
(port 9312).  Trying to use that port for normal Barbican
ReST calls resulted in the http 404.  This fix adds the
region to the devstack config and uses that on the filter
that Barbican gives to Tempest to find the correct
Barbican endpoint.

Change-Id: Ib6dd5aa79198463b5db2541d85de8e67b400212b
Closes-Bug: #1407767