Fix UnicodeDecodeError's in the functional tests
While running the functional tests some UnicodeDecodeErrors are
triggered since the logger can't properly print some binary characters
that are provided in both the requests (When PUTing a secret) and the
responses (when GETing a secret). So this catches those errors and
sets the logged string to be properly printed, which helps debugging
for errors.

Change-Id: Ia61e4fc0891775c15d725ac8b2d5178e31ec9bf1