Refactor order validation
The 'meta' parameter for order validation was being handled manually
instead of in the JSON schema; So this was fixed. On the other hand,
there was a rudimentary validation for both the type and the meta
parameters in the Order's PUT controller; this was removed, as it
created unnecessary complexity.

Change-Id: I6d6d3f942276bc45729cbf7cfa0d17635e1dc93d