Drop old namespace for some oslo libraries
The oslo team has decided to drop the "oslo." namespacing as it was
causing a lot of trouble. So, as described [1], the namespace was
dropped in favor of the "oslo_" imports that will be used.

Note that oslo.messaging still doesn't have a release supporting their
switch to oslo_messaging, but it will come in their next release. So I
will do the switch when they have it ready.

[1] http://blog.nemebean.com/content/whys-and-hows-oslo-namespace-change

Change-Id: I4743598668d9dde540275911467b94386bad0983