Add 'history' option to the migration script
This is an initial implementation of a 'history' option, which enables
you to see the available migration scripts and their appropriate
hashes. This is helpful when debugging migration errors, as it allows
you to easily get information about the scripts without having to see
the files (when using the verbose option).

Change-Id: Ib6bfa728689774586617180d3d43a493155ab0c2