Renamed outputted keys from base model
I renamed the fields deletion related fields that would potentially
get outputted when doing a "to_dict_fields" operation on any model
that inherits from the base model. Now, we do not display deleted
entities through the API. But at some point we could be using this for
auditing. So this change was done to promote consistency.

Change-Id: Ieb90005a1cb1b75bb6182366d774b219564b073f