Completely refactor PKCS11 plugin
In testing Rackspace discovered a race condition with the PyKCS11
implementation that caused corrupted data to be persisted to the
database in high load conditions. After investigating, it was decided
that an appropriate solution would be to replace PyKCS11 with cffi
bindings. This has the added advantage of removing the requirement for
the forked PyKCS11.

Change-Id: Id075ecb6b1890ba78952be6aaf2b0edba4148c43