Change exception when store plugin is misconfigured
If we try to get a plugin that supports the "retrieve" and "delete"
operations and the plugin is not there, is is probably due to a server
misconfiguration, since the plugin name is gotten from the metadata
from the database; And if it was there in the first place, it means
that it once was actually able to store that secret using a valid
plugin. Thus, a new exception is raised if this is the case.

Change-Id: I2be8b9dd17a7bd12f10e55945b09257fce616f3d
Closes-Bug: #1416075