Refactor exception handling in the app side
To simplify how the app handles the HTTP error codes, I refactored the
code so classes that should return a specific code should inherit from
BarbicanHTTPException. Also, it was taken into consideration that for
security reasons, we might need to log one thing and send another
message to the client that uses our API; So to address this, an
appropriate field was added to the base exception.

Change-Id: I4ef15a2287a651366a77d51de5a2a12c66bc0dc0