Fix "invalid credentials" error running functional tests
This fixes the "invalid credentials" error that you will get
if you create a clean barbican environment and try to run the
functional tests.  It is actually a workaround for a change
to tempest (see https://review.openstack.org/#/c/147778/) which
causes a config object to be created BEFORE we get a chance to
set the path to OUR config.  When we set the path to our config
it is essentially ignored (since the config has already been
created) so the credentials in the tempest config aren't the
ones we set - hence the invalid credentials error.

This isn't an issue in devstack gate because our config
file isn't used.

This fix ensures that we will get our path update done before
tempest builds its config so our values will be used by tempest
when the real config object is created.

Change-Id: If8381aff3233934a0c733e12d64605c2dbdfe60c