fix a use after free bug in the pkcs11 plugin
When generating the mechanism for encrypting/decrypting we allocated
the GCM struct and assigned it but then returned the mechanism. This
caused the variable holding ownership in Python to be out of scope and
the GCM struct was marked for garbage collection. A classic use after
free bug. Ugh. Credit for finding this goes to Rohit Bansal

This CR additionally creates namedtuples to hold references to the
data and more elegantly prevent this bug.

Change-Id: Ic105c569328d599ca2e989e4413381d658cefe02