Enforce secret and order IDs are valid UUIDs
The following CR will enforce a verification of the ID passed from the
URI to make sure that it's actually a valid UUID. So it wouldn't even
do a lookup to the database if the id is not valid.

Now, for testing purposes, I mock the call to the aforementioned
validation. This is because it's preferable to have human readable IDs
while testing instead of UUIDs, since this eases debugging.

Change-Id: I0e0231f2587324ccbc40e98d1b69254beb9a6eb3