Using a central secret store manager to remove lock
The old code is creating a new plugin manager for every
request which requires us to use a lock around get_manager()
to keep a HSM from barfing because connections haven't been
initialized. Moving to a single manager for the calls allows
for us to remove the lock and not have connection initialization
issues.

Change-Id: If7346e977c313f96ec920b98c41213934a07ff21