Fixing unable to retrieve req id bug
Properly returning the id if it was provided on the request. If not,
generate an id in the format expected by openstack services
and clients.

Context:
The original code in BaseContextMiddleware was copied from Glance
at the beginning of the project, but it has never worked because
we don't build our requests and responses like Glance does. As
a result, the old code has been logging a "Unabled to retrieve request
id from context" warning for every response. This change properly
retrieves the id (if there), or generates one if needed.

Closes-Bug: 1424221
Change-Id: I7e22474775de3c3f33968144d098bedbfc462e7a