Refactor Orders resource to use repository factories
Instead of passing around the repositories in the constructor of the
controller classes, now a factory function is called.

A Mixin class was added in the tests in order to patch the
aforementioned factory functions. Though, only the functions related
to the Orders resource were added to this Mixin class.

Change-Id: Ibf468f09e6837cb8658cfcd164f4f766997c132a