Refactor Secrets resource to use repository factories
Instead of passing around the repositories in the constructor of the
controller classes, now a factory function is called. The necessary
functions were added in order to mock these factory calls.

The Repository class was left as is, as the objects that use it will
be refactored in subsequent commits.

Change-Id: I667ff5ede667cfe5bd62f924aac6bd98936fef6f