Fixing test dependence on execution order
This fixes a problem I noticed where if someone uses the manager
singleton before the execution of this test, the test will fail because
it was instantiated with the wrong configuration values that cannot be
overridden. As a result, resetting the manager singleton global will
make sure that it's instantiated with the correct state for this test.

Change-Id: I19c7728a0e6b4cc14ab1084bb4da4c2be14c4a12