Fix for order obj expiration issue with SQLAlchemy
By default, SQLAlchemy expires object state on commit. In some cases
(usually testing), the new_order object expires causing the whole
transaction to fail because SQLAlchemy raises an exception saying that
the object isn't bound to a session. This is because we're trying to
access the id attribute on a expired object. This patch pulls a copy
of the order_id off before we commit to the db; which can expire the
new_order object.

Change-Id: I799f6a52c20304abb6e537f5cbe69926efe38ead