Splitting out remaining order tests from test_resources
This CR is a continuation of the refactor efforts to remove mocks
and split up test_resources.py. The plan is get these tests split out
and then come back around for a second pass and clean up the tests
more, reduce duplication, and add more quality assertions around the
data.

Change-Id: I82fe3c2e4e00dc32b3d58f662f8bf1809fa8aa8d