Deprecate old secret decrypt path both in code and docs
Since the introduction of appending of 'payload' to the path to
decrypt a secret was merged both in the server and the client. Only
thing missing was to change that in the documentation and to issue a
deprecation warning in the code. This CR implements that.

Note that this only issues a log warning in the server, and doesn't
disable the decryption of message through the "old way".

Change-Id: I64f371732747323c6f89bc26143ca2765c5ea237