Removed get_secret_type
The KeyAlgorithm class contained a method called get_secret_type. This
method attempted to guess the type of secret being stored based upon
the algorithm. This method only supported symmetric algorithms and
could not determine secret types for public/private key pairs because
they use the same algorithm.

This patch removes the method. This method is no longer needed now
that the API supports users specifying the secret type. See the
content-types blueprint for more information on that.

Change-Id: Ife2e3ac5f8cf702dab4b3cb5a64edf32ca21877f
Related to blueprint content-types